+++ app/concepts/tribunal_instance/stock/operation/retrieve_last_stock.rb
@@ -1,6 +1,6 @@
-module DataSource
+module TribunalInstance
   module Stock
-    module Task
+    module Operation
       class RetrieveLastStock < Trailblazer::Operation
 
         # ctx should have stocks_folder & stock_class
