+++ spec/concepts/tribunal_instance/stock/operation/retrieve_last_stock_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe DataSource::Stock::Task::RetrieveLastStock do
+describe TribunalInstance::Stock::Operation::RetrieveLastStock do
   class DummyStockClass < Stock; end
 
   context 'when no valid stocks are found' do
