+++ spec/concepts/tribunal_instance/stock/unit/operation/merge_greffe_zero_spec.rb
@@ -1,6 +1,6 @@
 require 'rails_helper'
 
-describe DataSource::Stock::TribunalInstance::Unit::Operation::MergeGreffeZero do
+describe TribunalInstance::Stock::Unit::Operation::MergeGreffeZero do
   subject do
     described_class.call(
       entreprise_code_greffe_0000: entreprise_code_greffe_0000,
